Alert, toast and loading

+++ src/app/favorites/favorites.page.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit, Inject } from '@angular/core';
 import { FavoriteService } from '../services/favorite.service';
 import { Dish } from '../shared/dish';
-import { IonItemSliding } from '@ionic/angular';
+import { IonItemSliding, ToastController, LoadingController, AlertController } from '@ionic/angular';
 
 @Component({
   selector: 'app-favorites',
@@ -15,6 +15,9 @@ export class FavoritesPage implements OnInit {
 
   constructor(
     private favoriteService: FavoriteService,
+    private toastController: ToastController,
+    private loadingController: LoadingController,
+    private alertController: AlertController,
     @Inject('BaseURL') public baseUrl
   ) { }
 
@@ -24,18 +27,79 @@ export class FavoritesPage implements OnInit {
 
   getFavoriteDishes() {
     this.favoriteService.getFavorites().subscribe(
-      favoriteDishes => this.favoriteDishes = favoriteDishes,
+      favoriteDishes => {
+        this.favoriteDishes = favoriteDishes;
+      },
       error => this.favoriteDishesErrorMessage = error
     );
   }
 
-  deleteFavorite(item: IonItemSliding, id: number) {
-    console.log(`Delete ${id}`);
-    this.favoriteService.deleteFavorite(id).subscribe(
-      favoriteDishes => this.favoriteDishes = favoriteDishes,
-      error => this.favoriteDishesErrorMessage = error
+  async deleteFavorite(item: IonItemSliding, id: number) {
+    const wantToDelete = await this.presentAlert(id);
+
+    console.log(`Want to delete ${wantToDelete}`);
+
+    if (wantToDelete) {
+      const loading = await this.presentLoading();
+
+      this.favoriteService.deleteFavorite(id).subscribe(
+        favoriteDishes => {
+          this.favoriteDishes = favoriteDishes;
+          this.presentToast(`Dish ${id} deleted successfully`);
+          loading.dismiss();
+        },
+        error => {
+          this.favoriteDishesErrorMessage = error;
+          loading.dismiss();
+        }
+      );
+      item.close();
+    }
+  }
+
+  async presentToast(message: string) {
+    const toastController = await this.toastController.create({
+      message: message,
+      duration: 3000
+    });
+    toastController.present();
+  }
+
+  async presentLoading() {
+    const loading = await this.loadingController.create({
+      message: 'Deleting ...'
+    });
+    await loading.present();
+    return loading;
+  }
+
+  async presentAlert(id: number) {
+    return new Promise<Boolean>(
+      async (resolve, reject) => {
+        const alert = await this.alertController.create({
+          header: 'Confirm delete',
+          message: `Do you want to delete Dish ${id}`,
+          buttons: [
+            {
+              text: 'Cancel',
+              role: 'cancel',
+              handler: () => {
+                console.log('cancel');
+                resolve(false);
+              }
+            },
+            {
+              text: 'Delete',
+              handler: () => {
+                console.log('delete');
+                resolve(true);
+              }
+            }
+          ]
+        });
+        alert.present();
+      }
     );
-    item.close();
   }
 
 }
